code style improvements
